fixed maze_ctor loading issue
we originally were serializing the maze_ctor as a dict, then @mivanit
changed it to save just the name. this caused issues. We now save
as a (different than before) dict, but the loading_fn handles
both cases.

Since the code for saving as a string was only up for a short time,
warning will be printed if loading a model with the maze_ctor as string
with the hope that we can remove that soon